fix a bug for get MGR member info
we must consisder the network partition scenario.
if s1,s2,s1 are group, and s1 is primary node, if there is
a disconnection between (s1) and (s2,s3) and network partition,
for majority (s2,s3), new primary node may elect (if s2), but the
old primary s1's state still primary, we can't tolerate this.