Fix bug: granting issue on remote node switch
The grant issue occurs when more than 1 remote node is down and a remote node switch is attempted. Because tc_do_grants_internal() builds a connection map for all remote nodes before doing grants on the to-become-master slave node, when more than 1 node is down, the connection map will fails to be built, thus resulting in an error. This is not desirable since we need to allow remote node switch even if multiple remote nodes are down.