New Systeam Variable 'tc_auto_fix_connections'
Desc:
The new sysvar allows user to control whether to auto-fix existing remote connections that are either killed or timed out. If false, an error will be raised to report the lost connection's info.

Also a bugfix:
Fix the problem that some TDBCTL commands can cause the server to crash because of NULL connections.